there is already an excellent chrome-web-extension called taborama (https://chrome.google.com/webstore/detail/taborama-group-your-tabs/flicfhphobgnnpfhieodmggibdhadhnh) in the chrome web store.
As firefox webextension are compatible with chrome-webextension ports
from of chrome extensions to firefox will increase, I think it's best to
use a different name.

Furthermore, my taborama actually does not have a panorama'ish view
(which was where it was derived from), so the name does not make sense
anyways.

Conex is short for container express which again is a type of shipping container and
therefore a good match to something that operates on containers and
provides a way to handle tabs and containers quickly.

fixes kesselborn/conex#46