Remember if user was in a non-contained container
fixes kesselborn/conex#58

This will open tabs non-contained if the user was in a non-contained tab
before. This introduces a race condition though:

If tab movemement is enabled, the user is in a contained tab and then opens tabs multiple tabs very
fast (as in: 2 per second), at some point of time, the new tab will be
opened in a non-contained tab. I don't think I can work around it but I
don't know whether this is a big deal: only happens on new tabs which
means no info should be leaked.